pull a container into a namespace
Pull an image provided as an argument to the program. We're only
interested in images for Linux/amd64 at the moment.

We setup a default namespace for containerd named `c2vm`. Images that
will be pulled by containerd will be stored inside that namespace.

Once the program is build it can be run like this:
```
; sudo ./c2vm -container docker.io/library/redis:latest
pulled docker.io/library/redis:latest (38667897 bytes)
```

And the image is indeed in the namespace:
```
; sudo ctr -n c2vm images ls -q
docker.io/library/redis:latest
```