extract layers to a mounted loop device
We create a loop device by pre-allocating space to a file on disk. This
file is then converted to an ext4 partition, which we mount to a
temporary path on the host.

Once this is done, we extract all the layers from the given container on
that mounted path. We also add a few extra files to the image
(`/etc/hosts` and `/etc/resolv.conf`).

When we're done extracting and writing, we run resize2fs in order to
resize the image to a more reasonable size.