inject a script for init
Once we will be ready to boot our VM, we will need an init script in
place. For this, we create a simple shell script which we populate using
the environment variables and command extracted from the container.

The init script is stored in /init.sh within the new image, and we will
configure the boot parameter to find it there.