boot the VM
The binary needs a few more arguments: the path to the firecracker
binary, the path to a linux kernel.

Using the image that was generated, we can now boot the VM with
firecracker. This will rely on a CNI configuration to create the
network, and will use the provided kernel to boot.